+++ Mail/HelperInterface.php
@@ -4,14 +4,20 @@ namespace Sulu\Bundle\FormBundle\Mail;
 
 interface HelperInterface
 {
+    const MAIL_RECEIVER_TO = 'to';
+    const MAIL_RECEIVER_CC = 'cc';
+    const MAIL_RECEIVER_BCC = 'bcc';
+
     /**
      * @param string $subject
      * @param string $body
-     * @param string $toMail
+     * @param string|array $toMail
      * @param string $fromMail
      * @param bool $html
      * @param string $replyTo
      * @param \SplFileInfo[] $attachments
+     * @param string|array $ccMail
+     * @param string|array $bccMail
      *
      * @return int
      */
@@ -22,6 +28,15 @@ interface HelperInterface
         $fromMail = null,
         $html = true,
         $replyTo = null,
-        $attachments = []
+        $attachments = [],
+        $ccMail = [],
+        $bccMail = []
     );
+
+    /**
+     * Returns an array for holding receivers divided by types.
+     *
+     * @return array
+     */
+    public function getReceiverTypes();
 }